ajuste en logout
