+++ db.json
@@ -174,6 +174,15 @@
       "tasks": [],
       "users": [],
       "createdAt": "2024-03-12T00:12:43.665Z"
+    },
+    {
+      "id": "arq-1710253855342522",
+      "projectName": "qwerqwer",
+      "description": "qwerqwer",
+      "limitDate": "2024-03-13",
+      "tasks": [],
+      "users": [],
+      "createdAt": "2024-03-12T14:30:55.343Z"
     }
   ],
   "tasks": [
@@ -204,7 +213,11 @@
       "password": "123123",
       "isLoged": false,
       "token": "",
-      "id": "a052"
+      "id": "a052",
+      "user": {
+        "isLoged": false,
+        "token": ""
+      }
     },
     {
       "id": "9aac",
@@ -223,7 +236,7 @@
       "tasks": [],
       "role": "MANAGER",
       "isLoged": true,
-      "token": "arq-1710251750439270arq-1710251750439442"
+      "token": "arq-1710254173827802arq-171025417382753"
     }
   ]
 }
