+++ src/layouts/home.layout.jsx
@@ -1,10 +1,12 @@
 /* eslint-disable react-refresh/only-export-components */
 import styled from "styled-components";
 import enviroment from "../../enviroment";
-import { Link, Outlet, useLoaderData, useLocation } from 'react-router-dom';
+import { Link, Outlet, useLoaderData, useLocation, useNavigate } from 'react-router-dom';
 import { Header } from '../compnents/header'
 import { isValidArray } from "../helpers/validators";
 import { generateUniqueId } from "../helpers/unique_id";
+import KeyboardArrowLeftIcon from '@mui/icons-material/KeyboardArrowLeft';
+import { Button } from "@mui/material";
 
 export async function loader() {
     const url = enviroment.api_url;
@@ -25,6 +27,11 @@ const Wrapper = styled.div`
 function HomeLayout() {
     const menuItems = useLoaderData();
     const location = useLocation();
+    const navigate = useNavigate();
+    const handlerBack = ev => {
+        ev.preventDefault();
+        navigate(-1);
+    }
     return (
         <Wrapper className="container bg-slate-300 md:min-h-screen">
             <Header className="shadow-md mb-4">
@@ -56,6 +63,9 @@ function HomeLayout() {
                 </nav>
             </Header>
             <main className="bg-slate-50 rounded-md shadow-md px-5 py-10 mx-3 min-h-screen">
+                <Button variant="text" onClick={handlerBack}>
+                    <KeyboardArrowLeftIcon /> Back
+                </Button>
                 <Outlet />
             </main>
         </Wrapper>
