+++ src/pages/auth/logout.page.jsx
@@ -1,5 +1,7 @@
-import { redirect } from "react-router-dom";
+/* eslint-disable react-hooks/exhaustive-deps */
+import { redirect, useNavigate } from "react-router-dom";
 import { logout } from "../../services/auth";
+import { useEffect } from "react";
 
 /* eslint-disable react-refresh/only-export-components */
 export async function loader() {
@@ -7,7 +9,6 @@ export async function loader() {
         const storage = localStorage.getItem('userId');
         if (!storage) return "";
         const res = await logout({ id: "a052" });
-        console.log({ res });
         res && localStorage.removeItem("userId");
         return redirect("/");
     } catch (e) {
@@ -15,6 +16,10 @@ export async function loader() {
     }
 }
 function LogoutPage() {
+    const navigate = useNavigate();
+    useEffect(() => {
+        navigate("/")
+    }, []);
     return (<div className="container p-10 m-10 text-center"> `/// Cargando ///` </div>);
 }
